move VSLIceConfiguration and VSLTurnConfiguration classes out of VSLAccountConfiguration.h file
