+++ Pod/Classes/Configurations/VSLIceConfiguration.h
@@ -1,0 +1,12 @@
+//
+//  VSLIceConfiguration.h
+//  VialerSIPLib
+//
+//  Created by Maciek Gierszewski on 25/10/2017.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface VSLIceConfiguration : NSObject
+@property (nonatomic, assign) BOOL enableIce;
+@end
