+++ Pod/Classes/Configurations/VSLIceConfiguration.m
@@ -1,0 +1,12 @@
+//
+//  VSLIceConfiguration.m
+//  VialerSIPLib
+//
+//  Created by Maciek Gierszewski on 25/10/2017.
+//
+
+#import "VSLIceConfiguration.h"
+
+@implementation VSLIceConfiguration
+
+@end
