+++ Pod/Classes/Configurations/VSLTurnConfiguration.m
@@ -1,0 +1,12 @@
+//
+//  VSLTurnConfiguration.m
+//  VialerSIPLib
+//
+//  Created by Maciek Gierszewski on 25/10/2017.
+//
+
+#import "VSLTurnConfiguration.h"
+
+@implementation VSLTurnConfiguration
+
+@end
